add authentication & authorization

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { DeleteResult } from 'typeorm';
+import { User } from 'src/auth/user.entity';
 
 
 @Injectable()
@@ -16,29 +17,29 @@ export class TasksService {
         private taskRepo: TaskRepository
     ) {}
 
-    getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepo.getTasks(filterDto);
+    getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepo.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepo.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const task = await this.taskRepo.findOne({where: {id, userId: user.id}});
         if(!task) throw new NotFoundException(`Task with id ${id} is not found.`);
         return task;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepo.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepo.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
+    async deleteTask(id: number, user: User): Promise<void> {
         // const task = await this.getTaskById(id);
-        const result = await this.taskRepo.delete(id);
+        const result = await this.taskRepo.delete({id, userId: user.id});
         console.log(result);
         if(result.affected === 0) throw new NotFoundException(`Task with id ${id} is not found.`);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
